chore: split project into engine and web packages
