+++ package.json
@@ -1,32 +1,10 @@
 {
-  "name": "kingdom-builder-engine",
-  "version": "0.1.0",
+  "name": "kingdom-builder",
   "private": true,
   "type": "module",
+  "workspaces": ["packages/*"],
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "preview": "vite preview",
-    "lint": "eslint . --ext .ts,.tsx",
-    "test": "npm run lint && vitest run"
-  },
-  "dependencies": {
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
-  },
-  "devDependencies": {
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "@typescript-eslint/eslint-plugin": "^7.14.1",
-    "@typescript-eslint/parser": "^7.14.1",
-    "autoprefixer": "^10.4.21",
-    "eslint": "^8.57.0",
-    "eslint-plugin-unused-imports": "^3.1.0",
-    "postcss": "^8.5.6",
-    "tailwindcss": "^3.4.17",
-    "typescript": "^5.5.4",
-    "vite": "^5.4.0",
-    "vitest": "^3.2.4"
+    "build": "npm run build --workspaces",
+    "test": "npm test --workspaces"
   }
 }
