+++ packages/engine/src/effects/add_resource.ts
