+++ packages/engine/src/effects/pay_resource.ts
