+++ packages/engine/tests/effects/till_land.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { createEngine, performAction, createActionRegistry } from '../../src/engine';
+import { createEngine, performAction, createActionRegistry } from '../../src/index.ts';
 
 describe('till_land effect', () => {
   it('increases land slots up to the maximum', () => {
