+++ packages/web/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "noEmit": true,
+    "composite": true
+  },
+  "include": ["src"],
+  "references": [
+    { "path": "../engine" }
+  ]
+}
