+++ tsconfig.json
@@ -1,17 +1,7 @@
 {
-  "compilerOptions": {
-    "target": "ES2022",
-    "useDefineForClassFields": true,
-    "lib": ["ES2022", "DOM", "DOM.Iterable"],
-    "module": "ESNext",
-    "skipLibCheck": true,
-    "moduleResolution": "Bundler",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
-    "jsx": "react-jsx",
-    "strict": true,
-    "types": ["vitest"]
-  },
-  "include": ["src", "tests"]
+  "files": [],
+  "references": [
+    { "path": "packages/engine" },
+    { "path": "packages/web" }
+  ]
 }
