--- tsconfig.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/app.tsx"],"version":"5.9.2"}
